photo_sharing_site_root/requirements.txt
absl-py==0.4.1
astor==0.7.1
bleach==1.5.0
certifi==2018.8.24
chardet==3.0.4
Django==2.1
django-appconf==1.0.2
django-compressor==2.2
djangorestframework==3.8.2
flake8==3.5.0
gast==0.2.0
grpcio==1.14.2
html5lib==0.9999999
idna==2.7
Markdown==2.6.11
mccabe==0.6.1
numpy==1.15.1
pika==0.12.0
Pillow==5.1.0
protobuf==3.6.1
pycodestyle==2.3.1
pyflakes==1.6.0
pytz==2018.5
rcssmin==1.0.6
requests==2.19.1
rjsmin==1.0.12
six==1.11.0
tensorboard==1.7.0
tensorflow==1.7.1
tensorflow-hub==0.1.1
termcolor==1.1.0
urllib3==1.23
Werkzeug==0.14.1
